update cyber grind (need to polish)
